feature: add login page style

+++ src/lib/utils/sizes.dart
@@ -13,9 +13,9 @@ class Sizes {
   static double getHalfHeight(BuildContext context) =>
       MediaQuery.of(context).size.height / 2;
 
-  static double getPercentWidth(BuildContext context, int percent) =>
+  static double getPercentWidth(BuildContext context, double percent) =>
       MediaQuery.of(context).size.width * (percent / 100);
 
-  static double getPercentHeight(BuildContext context, int percent) =>
+  static double getPercentHeight(BuildContext context, double percent) =>
       MediaQuery.of(context).size.height * (percent / 100);
 }
